Rename math.cc to geometry.cc
as that's really what it was about
