+++ src/input.cc
@@ -26,7 +26,7 @@
 #include <string>
 #include <vector>
 
-#include "libpolygon/elementary_geometry.h"
+#include "libpolygon/geometry.h"
 #include "libpolygon/globals.h"
 #include "libpolygon/number_field.h"
 #include "libpolygon/oedge.h"
