+++ src/libpolygon/geometry.h
