+++ src/psimplex.cc
@@ -24,7 +24,7 @@
 #include <list>
 #include <string>
 
-#include "libpolygon/elementary_geometry.h"
+#include "libpolygon/geometry.h"
 #include "libpolygon/globals.h"
 #include "libpolygon/my_ostream.h"
 #include "libpolygon/pface.h"
