+++ src/two_complex.cc
@@ -24,7 +24,7 @@
 #include <ostream>
 #include <vector>
 
-#include "libpolygon/elementary_geometry.h"
+#include "libpolygon/geometry.h"
 #include "libpolygon/globals.h"
 #include "libpolygon/number_field.h"
 #include "libpolygon/params.h"
